changed indices to long

+++ model.py
@@ -314,9 +314,9 @@ class DeepPairWiseWord(nn.Module):
 		sorted, indices=torch.sort(simCube[6].view(1,-1),descending=True)
 		record=[]
 		for indix in indices[0]:
-			pos1=torch.div(indix,simCube.size(2)).data
-			pos2=(indix-simCube.size(2)*pos1).data
-			if s1tag[pos1.long()]+s2tag[pos2.long()]<=0:
+			pos1=torch.div(indix,simCube.size(2)).data.long()
+			pos2=(indix-simCube.size(2)*pos1).data.long()
+			if s1tag[pos1]+s2tag[pos2]<=0:
 				s1tag[pos1]=1
 				s2tag[pos2]=1
 				record.append((pos1,pos2))
@@ -338,8 +338,8 @@ class DeepPairWiseWord(nn.Module):
 		sorted, indices = torch.sort(simCube[7].view(1, -1), descending=True)
 		counter=0
 		for indix in indices[0]:
-			pos1 = torch.div(indix, simCube.size(2)).data
-			pos2 = (indix-simCube.size(2)*pos1).data
+			pos1 = torch.div(indix, simCube.size(2)).data.long()
+			pos2 = (indix-simCube.size(2)*pos1).data.long()
 			if s1tag[pos1] + s2tag[pos2] <= 0:
 				counter+=1
 				if (pos1,pos2) in record:
